+++ backend/src/main/java/mouda/backend/auth/presentation/controller/swagger/AuthSwagger.java
@@ -10,9 +10,7 @@ import mouda.backend.auth.presentation.request.AppleOauthRequest;
 import mouda.backend.auth.presentation.request.GoogleOauthRequest;
 import mouda.backend.auth.presentation.request.OauthRequest;
 import mouda.backend.auth.presentation.response.LoginResponse;
-import mouda.backend.common.config.argumentresolver.LoginMember;
 import mouda.backend.common.response.RestResponse;
-import mouda.backend.member.domain.Member;
 
 public interface AuthSwagger {
 
@@ -33,11 +31,4 @@ public interface AuthSwagger {
 		@ApiResponse(responseCode = "200", description = "로그인 성공!"),
 	})
 	ResponseEntity<RestResponse<LoginResponse>> loginGoogleOauth(@RequestBody GoogleOauthRequest googleOauthRequest);
-
-	@Operation(summary = "회원 탈퇴", description = "로그인한 회원을 서비스에서 탈퇴 처리한다.")
-	@ApiResponses({
-		@ApiResponse(responseCode = "200", description = "회원 탈퇴 성공!"),
-	})
-	ResponseEntity<Void> withdraw(@LoginMember Member member);
-
 }
