+++ backend/src/main/java/mouda/backend/member/business/MemberService.java
@@ -7,6 +7,7 @@ import lombok.RequiredArgsConstructor;
 import mouda.backend.auth.implement.jwt.AccessTokenProvider;
 import mouda.backend.member.domain.Member;
 import mouda.backend.member.implement.MemberFinder;
+import mouda.backend.member.implement.MemberWriter;
 
 @Service
 @Transactional
@@ -15,6 +16,7 @@ public class MemberService {
 
 	private final AccessTokenProvider accessTokenProvider;
 	private final MemberFinder memberFinder;
+	private final MemberWriter memberWriter;
 
 	public Member findMember(String token) {
 		String socialId = accessTokenProvider.extractSocialId(token);
@@ -23,5 +25,9 @@ public class MemberService {
 
 	public void checkAuthentication(String token) {
 		accessTokenProvider.validateToken(token);
+	}
+
+	public void withdraw(Member member) {
+		memberWriter.withdraw(member);
 	}
 }
